[refactor] RedisService objectMapper 적용
